types: move StructuralType to structural.rs
Code move only.
